Refactor Apps component to integrate EditChatAppSheet and EditCrewAppSheet
- Removed the previous EditAppSheet implementation and replaced it with conditional rendering for EditChatAppSheet and EditCrewAppSheet based on the selected app type.
- Enhanced data fetching logic to include tasks and variables alongside tools, models, and agents.
- Updated state management to accommodate new task and variable choices.
- Improved overall structure and readability of the Apps component.
