+++ webapp/src/pages/[resourceSlug]/apps.tsx
@@ -50,6 +50,8 @@ import {
 } from 'modules/components/ui/sheet';
 import { Model } from 'db/model';
 import { Agent } from 'struct/agent';
+import EditChatAppSheet from 'components/apps/EditAppSheet';
+import EditCrewAppSheet from 'components/apps/EditCrewAppSheet';
 
 const DeleteDialog = ({
 	openDeleteDialog,
@@ -106,86 +108,6 @@ const DeleteDialog = ({
 	);
 };
 
-const EditAppSheet = ({
-	open,
-	setOpen,
-	selectedApp,
-	onAppUpdate,
-	toolChoices,
-	modelChoices,
-	agentChoices,
-	fetchFormData // Add this prop
-}: {
-	open: boolean;
-	setOpen: (open: boolean) => void;
-	selectedApp: App;
-	onAppUpdate: () => void;
-	toolChoices: any[];
-	modelChoices: any[];
-	agentChoices: any[];
-	fetchFormData: () => Promise<void>; // Add this type
-}) => {
-	// Add default values for missing properties
-	const enrichedApp = {
-		...selectedApp,
-		chatAppConfig: {
-			...selectedApp?.chatAppConfig,
-			maxMessages: selectedApp?.chatAppConfig?.maxMessages || 30
-		},
-		sharingConfig: {
-			...selectedApp?.sharingConfig,
-			mode: selectedApp?.sharingConfig?.mode || SharingMode.TEAM,
-			permissions: selectedApp?.sharingConfig?.permissions || {}
-		},
-		description: selectedApp?.description || '',
-		name: selectedApp?.name || 'Untitled Chat App',
-		type: selectedApp?.type || AppType.CHAT
-	};
-
-	return (
-		<Sheet
-			open={open}
-			onOpenChange={open => {
-				if (!open) {
-					// Force a slight delay before resetting pointer events
-					setTimeout(() => {
-						document.body.style.pointerEvents = 'auto';
-						document.body.style.cursor = 'auto';
-					}, 100);
-				}
-				setOpen(open);
-			}}>
-			<SheetContent
-				size='xl'
-				className='w-full overflow-visible'
-				style={{
-					pointerEvents: 'auto',
-					overflow: 'visible'
-				}}>
-				<SheetHeader>
-					<SheetTitle>Edit App</SheetTitle>
-					<SheetDescription>Modify your app settings and configuration.</SheetDescription>
-				</SheetHeader>
-				<div className='mt-4 overflow-auto max-h-[80vh]'>
-					<ChatAppForm2
-						fetchFormData={fetchFormData}
-						app={enrichedApp}
-						editing={true}
-						callback={() => {
-							setOpen(false);
-							onAppUpdate();
-						}}
-						toolChoices={toolChoices} // Pass these from parent if available
-						modelChoices={modelChoices} // Pass these from parent if available
-						agentChoices={agentChoices} // Pass these from parent if available
-						whiteListSharingChoices={Object.values(enrichedApp.sharingConfig.permissions || {})}
-					/>
-				</div>
-			</SheetContent>
-		</Sheet>
-	);
-};
-
 export default function Apps(props) {
 	const { agents } = useAgentStore();
 	const [openDeleteDialog, setOpenDeleteDialog] = useState<boolean>(false);
@@ -225,15 +147,19 @@ export default function Apps(props) {
 	const [toolChoices, setToolChoices] = useState<any[]>([]);
 	const [modelChoices, setModelChoices] = useState<Model[]>([]);
 	const [agentChoices, setAgentChoices] = useState<Agent[]>([]);
+	const [taskChoices, setTaskChoices] = useState<any[]>([]);
+	const [variableChoices, setVariableChoices] = useState<any[]>([]);
 
 	// Add function to fetch all required data
 	const fetchFormData = async () => {
 		try {
-			// Add your API calls here to fetch tools, models, and agents
-			const [toolsRes, modelsRes, agentsRes] = await Promise.all([
+			// Add your API calls here to fetch tools, models, agents, tasks, and variables
+			const [toolsRes, modelsRes, agentsRes, tasksRes, variablesRes] = await Promise.all([
 				API.getTools({ resourceSlug, _csrf: csrf }, null, setError, router),
 				API.getModels({ resourceSlug, _csrf: csrf }, null, setError, router),
-				API.getAgents({ resourceSlug, _csrf: csrf }, null, setError, router)
+				API.getAgents({ resourceSlug, _csrf: csrf }, null, setError, router),
+				API.getTasks({ resourceSlug, _csrf: csrf }, null, setError, router),
+				API.getVariables({ resourceSlug, _csrf: csrf }, null, setError, router)
 			]);
 
 			// Filter out embedding models - we only want chat/completion models
@@ -243,6 +169,8 @@ export default function Apps(props) {
 			setToolChoices(toolsRes?.tools || toolsRes?.data || []);
 			setModelChoices(chatModels);
 			setAgentChoices(agentsRes?.agents || []);
+			setTaskChoices(tasksRes?.tasks || []);
+			setVariableChoices(variablesRes?.variables || []);
 		} catch (error) {
 			console.error('Error loading data:', error);
 			toast.error('Error loading data');
@@ -250,6 +178,8 @@ export default function Apps(props) {
 			setToolChoices([]);
 			setModelChoices([]);
 			setAgentChoices([]);
+			setTaskChoices([]);
+			setVariableChoices([]);
 		}
 	};
 
@@ -347,16 +277,32 @@ export default function Apps(props) {
 				setOpenDeleteDialog={setOpenDeleteDialog}
 				onDelete={() => deleteApp(selectedAgentId)}
 			/>
-			<EditAppSheet
-				open={openEditSheet}
-				setOpen={setOpenEditSheet}
-				selectedApp={selectedApp}
-				onAppUpdate={fetchApps}
-				toolChoices={toolChoices}
-				modelChoices={modelChoices}
-				agentChoices={agentChoices}
-				fetchFormData={fetchFormData} // Add this prop
-			/>
+			{selectedApp?.type === 'crew' ? (
+				<EditCrewAppSheet
+					open={openEditSheet}
+					setOpen={setOpenEditSheet}
+					selectedApp={selectedApp}
+					onAppUpdate={fetchApps}
+					toolChoices={toolChoices}
+					modelChoices={modelChoices}
+					agentChoices={agentChoices}
+					taskChoices={taskChoices}
+					variableChoices={variableChoices}
+					fetchFormData={fetchFormData}
+				/>
+			) : (
+				<EditChatAppSheet
+					open={openEditSheet}
+					setOpen={setOpenEditSheet}
+					selectedApp={selectedApp}
+					onAppUpdate={fetchApps}
+					toolChoices={toolChoices}
+					modelChoices={modelChoices}
+					agentChoices={agentChoices}
+					fetchFormData={fetchFormData}
+				/>
+			)}
+
 			<main className='text-foreground flex flex-col gap-2'>
 				<section className='flex items-center justify-between mb-4'>
 					<h4 className='text-gray-900 font-semibold text-2xl'>Apps</h4>
